+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Palindromic decomposition".
 - Tests for "Battleship".
 - Tests for "Texas holdem".
+- Tests for "Solid integer".
 
 ## [1.15.0] - 2022-11-30
 ### Added